Struct ruplicity::backend::local::LocalBackend

pub struct LocalBackend {
    // some fields omitted
}

Backend for some directory in the local file system.

Methods

impl LocalBackend

fn new<P: AsRef<Path>>(path: P) -> Self

Creates a new local backend for the given directory.