Module ruplicity::collections

Operations on backup files.

This sub-module provides information about a backup, by looking at the files present in a backup directory.

Structs

BackupChain
Contains information about a backup chain.

BackupSet
Information about the files which make a backup snapshot.

Collections
General information about a backup.

SignatureChain
Contains information about signatures in a backup chain.

SignatureFile
Information about a signature file.

Type Definitions

BackupSetIter
Iterator over BackupSets.

ChainIter
Iterator over some kind of chain.

SignatureFileIter
Iterator over SignatureFiles.