Struct ruplicity::collections::BackupSet

pub struct BackupSet {
    // some fields omitted
}

Information about the files which make a backup snapshot.

This struct contains paths for the manifest and volumes, and the backup format.

Methods

impl BackupSet

fn is_complete(&self) -> bool

Returns whether the set is complete.

The corresponding backup has not been stopped before completion.

fn is_partial(&self) -> bool

Returns whether the set is partial.

This could happen when a backup is stopped before the completion.

fn start_time(&self) -> Timespec

Returns the creation time of the parent set.

The parent set depends on the type of this set. If it is incremental, the parent set is the previous one. If it is full, there is no parent set, so the time returned is the creation time for this set. Otherwise the creation time of this set is returned. This time, combined with end_time represent the time span handled by this set.

fn end_time(&self) -> Timespec

Returns the creation time of this set.

If this set is a full backup, this time is the same as the start time. This time, combined with start_time represent the time span handled by this set.

fn is_compressed(&self) -> bool

Returns whether the set is compressed.

fn is_encrypted(&self) -> bool

Returns whether the set is encrypted.

fn manifest_path(&self) -> &str

Returns the path of the manifest file for the set.

fn volume_path(&self, volume_num: usize) -> Option<&str>

Returns the path of the given volume.

fn num_volumes(&self) -> usize

Returns the number of volumes in the set.

fn is_full(&self) -> bool

Returns whether the set is a full backup.

fn is_incremental(&self) -> bool

Returns whether the set is a full backup.

fn is_same_set(&self, pr: &Info) -> bool

Returns whether the given file belongs to the same backup set, by looking at timestamps.