Struct ruplicity::collections::Collections

pub struct Collections {
    // some fields omitted
}

General information about a backup.

Determines the status of a backup by looking at the files present in the backup folder. No backup archive is opened in this process. Thanks to that, performances are great; however no validation is performed on backup files.

Methods

impl Collections

fn new() -> Self

Creates a new empty collection.

fn from_filenames<I>(filenames: I) -> Self where I: IntoIterator, I::Item: AsRef<Path>

Creates a collection, starting from a list of file names.

The given file names are not opened for validation. Information is collected based solely by the names themselves.

Examples

use ruplicity::collections::Collections;

let names = vec!["duplicity-full.20150617T182545Z.manifest",
                 "duplicity-full.20150617T182545Z.vol1.difftar.gz",
                 "duplicity-full-signatures.20150617T182545Z.sigtar.gz",
                 "duplicity-inc.20150617T182545Z.to.20150617T182629Z.manifest",
                 "duplicity-inc.20150617T182545Z.to.20150617T182629Z.vol1.difftar.gz",
                 "duplicity-inc.20150617T182629Z.to.20150617T182650Z.manifest"];
let collections = Collections::from_filenames(&names);
assert_eq!(collections.backup_chains().count(), 1);
assert_eq!(collections.signature_chains().count(), 1);

fn backup_chains(&self) -> ChainIter<BackupChain>

Returns the backup chains.

Each backup chain should be coupled with a signature chain. They can be matched because they are both in chronological order.

fn signature_chains(&self) -> ChainIter<SignatureChain>

Returns the signature chains.

Each signature chain should be coupled with a backup chain. They can be matched because they are both in chronological order.

fn num_snapshots(&self) -> usize

Returns the total number of snapshots.