Struct ruplicity::collections::SignatureChain

pub struct SignatureChain {
    // some fields omitted
}

Contains information about signatures in a backup chain.

See the docs for BackupChain.

Methods

impl SignatureChain

fn new(fname: &str, pr: &Info) -> Self

Create a new SignatureChain starting from a full signature.

fn from_filename_info(fname_info: &FileNameInfo) -> Self

Create a new signature chain from a FileNameInfo instance.

fn add_new_sig(&mut self, fname: &FileNameInfo) -> bool

Adds the given incremental signature to the signature chain if possible, returns false otherwise.

fn full_signature(&self) -> &SignatureFile

The file name of the full signature chain.

fn inc_signatures(&self) -> SignatureFileIter

A list of file names for incremental signatures.

fn start_time(&self) -> Timespec

Returns the time of the first signature in the chain.

fn end_time(&self) -> Timespec

Returns the time of the last signature in the chain.