Struct ruplicity::manifest::Manifest

pub struct Manifest {
    // some fields omitted
}

Manifest file info.

Methods

impl Manifest

fn parse<R: BufRead>(m: &mut R) -> Result<Self, ParseError>

Parses a stream to get a manifest.

fn hostname(&self) -> &str

The hostname produced the backup.

fn local_dir(&self) -> Option<&Path>

The original backup root path.

fn last_volume_index(&self) -> usize

The index of the last volume.

fn volume(&self, num: usize) -> Option<&Volume>

Returns the volume corresponding to the given index if present.

Note that volumes are counting starting from one, so the last volume number is equal to the number of volumes. If no volume corresponds to the given number, None is returned.

fn first_volume_of_path(&self, path: &[u8]) -> Option<usize>

Returns the index of the first volume containing the given path, if present.

The given path is represented with a byte array, because:

duplicity supports non-UTF8 paths;

under Windows Path is not allowed to contain non-UTF8 sequences.

fn last_volume_of_path(&self, path: &[u8]) -> Option<usize>

Returns the index of the last volume containing the given path, if present.

The given path is represented with a byte array, because:

duplicity supports non-UTF8 paths;

under Windows Path is not allowed to contain non-UTF8 sequences.