Struct ruplicity::manifest::Volume

pub struct Volume {
    // some fields omitted
}

Volume info.

Methods

impl Volume

fn start_path(&self) -> Option<&Path>

Returns the first path handled by this volume.

Note that the path can be None under Windows, in case it is non-UTF8. Use start_path_bytes if you need the byte array representing the path. This function never fails under *nix systems.

fn end_path(&self) -> Option<&Path>

Returns the last path handled by this volume.

Note that the path can be None under Windows, in case it is non-UTF8. Use start_path_bytes if you need the byte array representing the path. This function never fails under *nix systems.

fn start_path_bytes(&self) -> &[u8]

Returns the first path handled by this volume, represented as a byte array.

fn end_path_bytes(&self) -> &[u8]

Returns the first path handled by this volume, represented as a byte array.

fn start_block(&self) -> Option<usize>

Returns the number of the starting block of the first path.

If the first path is divided in multiple blocks and this volume does not start with the first block of that path, this function returns that block number.

fn end_block(&self) -> Option<usize>

Returns the number of the last block of the last path.

If the last path is divided in multiple blocks and this volume does not ends with the last block of that path, this function returns that block number.

fn hash_type(&self) -> &str

Returns a string representing the hash type of this volume.

fn hash(&self) -> &[u8]

Returns the hash value of this volume.