Struct ruplicity::signatures::Chain

pub struct Chain {
    // some fields omitted
}

Stores information about paths in a backup chain.

The information is reused among different snapshots if possible.

Methods

impl Chain

fn new() -> Self

Builds a new empty signature chain.

fn from_sigchain<B: Backend>(coll: &SignatureChain, backend: &B) -> Result<Self>

Opens a signature chain from signature chain files, by using a backend.

The given signature chain file names are read by using the given backend, to build the corresponding Chain instance.

fn snapshots(&self) -> Snapshots

Returns the snapshots present in the signature chain.