Struct ruplicity::signatures::Entry

pub struct Entry<'a> {
    // some fields omitted
}

Information about an entry inside a backup snapshot.

This could be a file, a directory, a link, etc.

Methods

impl<'a> Entry<'a>

fn path(&self) -> Option<&'a Path>

Returns the full path of the entry.

The path could be None if it is not UTF-8 representable under Windows. In that case use the byte representation with path_bytes.

fn path_bytes(&self) -> &'a [u8]

Returns the full path of the entry in bytes.

fn userid(&self) -> Option<u32>

Returns the value of the owner's user ID field.

fn groupid(&self) -> Option<u32>

Returns the value of the group's user ID field.

fn mode(&self) -> Option<u32>

Returns the mode bits for this file.

fn username(&self) -> Option<&'a str>

Returns the name of the owner user.

fn groupname(&self) -> Option<&'a str>

Returns the name of the group.

fn mtime(&self) -> Timespec

Returns the last modification time.

fn size_hint(&self) -> Option<(usize, usize)>

Returns a lower and upper bound in bytes on the entry size.

Note that for directories, this returns a size of zero, even if on Linux directories are often considered to have a 4096 bytes size.

fn entry_type(&self) -> EntryType

Returns the type of the entry.

fn linked_path(&self) -> Option<&'a Path>

Returns the path that this entry points to.

This will return some path only if this entry is a symbolic link.