Struct ruplicity::Backup

pub struct Backup<B> {
    // some fields omitted
}

A top level representation of a duplicity backup.

Methods

impl<B: Backend> Backup<B>

fn new(backend: B) -> Result<Self>

Opens an existig backup by using the given backend.

Errors

This function will return an error whenever the backend returns an error in a file operation. If the backend can't provide access to backup files, because they are unavailable or non-existing, an empty backup could be returned.

Examples

use ruplicity::Backup;
use ruplicity::backend::local::LocalBackend;

// use the local backend to open a path in the file system containing a backup
let backend = LocalBackend::new("tests/backups/single_vol");
let backup = Backup::new(backend).unwrap();
println!("Got backup with {} snapshots!", backup.snapshots().unwrap().into_iter().count());

fn snapshots(&self) -> Result<Snapshots>

Constructs an iterator over the snapshots currently present in this backup.

fn into_inner(self) -> B

Unwraps this backup and returns the inner backend.