Struct ruplicity::Snapshot

pub struct Snapshot<'a> {
    // some fields omitted
}

A snapshot in a backup.

Methods

impl<'a> Snapshot<'a>

fn time(&self) -> Timespec

Returns the time in which the snapshot has been taken.

fn is_full(&self) -> bool

Returns whether the snapshot is a full backup.

A full snapshot does not depend on previous snapshots.

fn is_incremental(&self) -> bool

Returns whether the snapshot is an incremental backup.

An incremental snapshot depends on all the previous incremental snapshots and the first previous full snapshot. This set of dependent snapshots is called "chain".

fn num_volumes(&self) -> usize

Returns the number of volumes contained in the snapshot.

fn as_backup_set(&self) -> &'a BackupSet

Returns the low level representation of the snapshot.

fn entries(&self) -> Result<SnapshotEntries>

Returns the files and directories present in the snapshot.

Be aware that using this functionality means that all the signature files in the current backup chain must be loaded, and this could take some time, depending on the file access provided by the backend and the signatures size.

fn manifest(&self) -> Result<ManifestRef<'a>, ParseError>

Returns the manifest for this snapshot.

The relative manifest file is read on demand and cached for subsequent uses.