Trait ruplicity::timefmt::TimeDisplay

pub trait TimeDisplay {
    type D: Display;
    fn into_local_display(self) -> Self::D;
    fn into_utc_display(self) -> Self::D;
}

Trait that allows to display a time into a local or UTC timezone.

Associated Types

type D: Display

The displayable type

Required Methods

fn into_local_display(self) -> Self::D

Turns self into a displayable type that when displayed uses the local time zone.

fn into_utc_display(self) -> Self::D

Turns self into a displayable type that when displayed uses the UTC time zone.